Deep-Learning/Knowledge-Distillation/requirements.txt
scipy==1.0.0
numpy==1.14.0
Pillow==8.1.1
tabulate==0.8.2
# tensorflow==1.7.0rc0
tensorflow==1.7.0
# torch==0.3.0.post4
torch
torchvision==0.2.0
tqdm==4.19.8
torchnet
